【统计分数】分页bug修改，计算均分问题解决！

+++ src/main/java/com/example/academic_affairs_management_system/controller/dto/AdminPack/Score.java
@@ -16,6 +16,7 @@ public class Score {
     private Integer testScore;
     private Integer score;
     private Integer ratio;
-    private Double finalScore;
+    private int finalScore;
     private Double scorePoint;
+
 }
